(parser) Add C++-based `StringTokenTypeDictionary`
...and use it for one of the fields in the `Scanner` class. This is a
small step towards #505.

Using a 3rd party library for now since I ended up with some weird
issues when using `std::unordered_map`. The problem turned out to be the
constructor not having been executed, because the CppSharp-generated C#
bindings for the C++ code wasn't up-to-date. This would lead to
uninitialized fields, causing errors when the code tried to use the map.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/577