(compiler) Replace `PerlangClass` with `IPerlangClass` interface
This is a simplification. I realized that this class shouldn't really be
needed, since it just wraps data which is already available in
`Stmt.Class`. For simplicity, we can just use `Stmt.Class` as-is
instead.

But, it still feels kind of nice to avoid having to depend on a very
particular detail in the syntax tree, namely the specific properties in
the `Stmt.Class` class. Given that we already referred to the
`PerlangClass` in various places, replacing these with a new
`IPerlangClass` interface and letting this class be implemented by
`Stmt.Class` feels like the best of both worlds.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/581