(compiler) Add support for immutable fields
The nice part about this is the syntax. While not claiming it to be
perfect in any way, I like the idea of "reversing" the const keyword
from other languages - let `const` be the default and `mutable` be the
slightly more awkward, opt-in. The idea is of course to encourage people
to design systems with an immutability-first mindset.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/583