(compiler) Remove empty statements in generated C++ code
This has been a mild annoyance for a while, but we haven't gotten around
to fixing it until now. Certain constructs have been emitting semicolons
on a line of their own in the generated C++ code, with or without
indentation.

This turned out to be two bugs in fact:

- Empty expressions would emit a semicolon with proper indentation. It
  would look like a real statement; it just didn't have any content.

- `new SomeClass()`-expressions would generate an extra semicolon
  without indentation.

Both of those are now fixed in this commit.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/585