(compiler) Add `extern` keyword for importing native functions/classes
This is an important piece of the puzzle for being able to rewrite our
`Scanner` class in pure Perlang goodness. To be able to do that in
practice, we'll need to interact with native C++ code that we write to
support the Perlang code (because of language features currently missing
in Perlang, like generics, full array support, etc).

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/586