(compiler) Remove `ARGV` superglobal and superglobals concept altogether
This was useful when we were an interpreter, and had its purpose. We
will still need a concept for accessing the command line variables, both
for `fun main()`-style programs and for programs using top-level
statements (our historical "mode of operation" when we had a REPL and
were more script-oriented), but for now, it:

1. does not work, and
2. forces us to create dummy values for certain `Token` properties
3. does not have any tests that have been executed for more than a year
  (2023-11-02).

Let's kill this with 🔥 for now. We might reintroduce something like
this again at some point, but let's not keep it lying around all the way
until then.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/587