(stdlib) Add more `print()` overloads
This is work towards being able to disable implicit overloads for the
method, but regretfully it's hard to take it all the way at the moment
because printing of Perlang `enum`s rely on implicit conversions. Either
way, committing this as a step towards hopefully completing this at some
point.

(Originally started as part of
https://gitlab.perlang.org/perlang/perlang/-/merge_requests/589)

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/590