(parser) Fix error handling for `char` literals
We didn't have any tests for the error cases, and we had bugs in how the
terminating single quotes was handled/consumed in case of errors. This
commit adds the missing tests. This ensures that the expected errors are
emitted (which was already partially implemented), and also that only
the expected number of errors are emitted for each error scenario. This
should provide a much better error message experience to the end user.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/591