(perlang_cli) Include short git commit ID in `perlang -v` output
This was actually inspired by a feature I implemented at work, where I
included the commit ID in the Semantic Version. Adding it there means
it'll be included in the `perlang -v` output`, essentially meaning that 
you'll hardly ever need to use `perlang -V` (show detailed version
information), at least not for snapshot versions.

Thinking loudly, this does actually mean that the commit ID will be
included in non-snapshot releases too. This may or may not be desirable;
we'll see if we want to revisit that going forward.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/593