(compiler) Include correct file name in compilation errors
The feature in !588 added the file name to our compilation errors, but
that only got us halfway. We still had code in
`CompileAndAssembleFile()` which would use a too simplistic approach
(added in !508) for compiling multiple files, where we simply
concatenated them all into one big string and passed into the compiler.
This will of course make it impossible to provide any form of reasonable
file-based compilation errors. This commit fixes that.

References:

- !508: https://gitlab.perlang.org/perlang/perlang/-/merge_requests/508
- !588: https://gitlab.perlang.org/perlang/perlang/-/merge_requests/588

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/595