(perlang_cli) Minor formatting cleanup
I changed the brace style in 508a570e099be6112bec5e869e020c9674c07415,
but now when working on a later change I realize that I probably used
the previous style on purpose. We should formalize this into some form
of "Perlang Style Guide" eventually, to make it easier both for myself
and others. ;)

For now, let's just state that the idea is:

- Braces on separate lines when opening a new method body
- Braces at end of statement when opening a block inside a method (for
  example, in a `for` loop or an `if` condition).

In other words, pretty much a modern form of the K&R style used in "The
C Programming Language".