(tests) Include original message in `EvalException`
Previously, the user has been forced to inspect the `InnerException`
(which is typically logged afterwards, by the test runner) in case
something goes wrong. This changes the logic so that the original
`e.Message` gets included in the newly thrown `EvalException` as well.

Also, the process output is changed to only be conditionally included.
If it's empty, there's little point in including it.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/602