(compiler) Add support for calling `get_type()` method
...which returns something similar to the `GetType()` method in .NET,
albeit much more simplistic. Also added support for `print()`:ing such
variables, so we can get our previous tests working again.

(This used to work before, so this is perhaps more bringing the
native-code-based compiler up to speed with "what we had", rather than
making new innovations in the language, but it's perhaps a bit
borderline)

For now, I think the working assumption is that all types of objects
supports a `get_type()` method, regardless of whether they are value
types (like `int` or `long`) or reference types (like `string` or
user-defined classes). On the C++ level, we will "emulate" it for
built-in value types for now, making it represent something like their
"boxed counterpart" (`perlang.Int32`) if/when we start going for an
auto-boxing model like C# and Java.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/605