(compiler) Rewrite `Func` to use `ITypeHandler` interface
Similarly as with
https://gitlab.perlang.org/perlang/perlang/-/merge_requests/572, this
changes from an FP style to a more OOP-oriented approach. In this case,
the change is also a preparation for a new function being added (which
retrieves a type by name, in contrast to adding/registering it), so the
interface makes even more sense in this case since it'll mean one
parameter to pass around instead of two (or more, if/when we start
adding even more similar methods).

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/606