(compiler) Remove native global class registration
This was useful while we were yet a .NET-based interpreter. We allowed a
handful of predefined classes to be registered as top-level global
classes, accessible to the Perlang world, to enable things like
Base64.encode("foo") and ARGV.pop() to work (the latter depending on a
"superglobal" concept somewhat similar to PHP, I think).

This is not yet supported in compiled mode, which is now (since 0.5.0)
the only supported mode. We still have some tests depending on these
classes to exist, but I think we can support them while dropping some of
the legacy code. The legacy code somewhat blocks new development coming
up, which is why this was actualized in the first place.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/607