(language) Support returning `this` from a method
This turned out a to be a bit non-trivial, because we expect the
instance of a class to be wrapped in `std::shared_ptr`. Now, how can you
_from the wrapped object_ get a hold of the shared pointer wrapping it?
This felt a bit like an impossible puzzle to solve. Luckily, both the
Boost library and eventually the C++ standard has us covered:

* https://www.boost.org/doc/libs/1_36_0/libs/smart_ptr/enable_shared_from_this.html
* https://en.cppreference.com/w/cpp/memory/enable_shared_from_this.html

(Implementing this was greatly helped by this and perhaps other SO
posts: https://stackoverflow.com/a/11711094/227779 🙏)

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/608