(language) Add `UTF16String` type
This change adds an initial version of an `UTF16String` type, as well as
an `as_utf16()` method in `UTF8String` for converting a UTF-8 string to
the corresponding UTF-16 string. The stdlib functionality for printing
strings have also been amended so that `UTF16String`s are printable (by
converting them to UTF-8 first).

Related to #370.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/620