(language) Add `length` property to `ASCIIString` and `UTF16String`
Interestingly enough, it didn't make sense to add this to the `String`
class, which does make sense if you think about the fact that this can
be either an `ASCIIString` or a `UTF8String`. Adding it to only the
fixed-width string classes for now, since that's where the initial need
is.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/624