(tests) Fix memory leak in integration test
Found by coincidence when working on other code; I realized that the
array allocated in the `StringHashSet::values_wrapper()` method would
never be freed.

An interesting aspect of this is that this hasn't been caught by our
Valgrind safeguards. I *think* that the tests exercising this method are
executed via Valgrind in fact, but it still doesn't seem to be detected
- perhaps because of the Valgrind configuation we use? Either way, this
should of course be fixed.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/632