(language) Add `object` data type
The `perlang::Object` class already existed on the C++ side, so we were
able to utilize that class for this. However, it was previously not the
base class for `perlang::String` and others, so this was also fixed now.

Objects of type `object` are not very usable right now, since we don't
have any form of typecasting mechanism (and also no reflection that
would enable some form of dynamic dispatch). They can be `print()`ed,
and stored in instance fields and variables, but that's pretty much it
right now.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/634