(general) Change license to LGPLv2.1
Note that this only relates to the compiler; the standard library is
still liberally licensed under the MIT license.

I think this strikes the balance between "preserve freedom for our
users" and "avoid unnecessary restrictions" quite well. The standard
library is more "generic" in nature and (at least at the moment) people
might want to copy content from it into their programs quite freely,
regardless of their license. This can perhaps be expected going forward
as well.

Now the *compiler* itself is a slightly different beast. I really don't
want anyone to come in and do a "non-free"/"closed source" fork of the
Perlang compiler. For the time being, I want to retain control over it.
The LGPL has strong guarantees making sure that if anyone ever wants to
fork it and distribute it, they will have to share their modifications
with us as well.

The choice of LGPL instead of GPL: I believe this will serve our needs
better, because I'm envisioning Perlang plugins/extensions being
developed for both the JetBrains and/or VS Code family of text editors
in the future. I want the compiler to be embeddable for such use cases,
even in proprietary versions of these programs like CLion and VS Code.
The LGPL license makes this possible, whereas GPL licensing would be
problematic since the plugin and the IDE can be considered a form of
"combined work", breaking the license. Now, we could always provide a
linking exception for this scenario but just using LGPLv2.1 as-is is
simpler.

And finally: Why version 2.1 and not version 3.0? For now, I'm not
convinced that the reciprocal/anti-patent-litigation aproach taken by
LGPLv3 (and GPLv3) is the way to go. I want the licensing to be "less
restrictive" in that sense. If I change my mind on this standpoint, I
might be willing to consider LGPLv3 but for now, I prefer a "Linux
approach" and sticking on the v2 series of these licenses.

Closes #419

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/635