(stdlib) Add `perlang::Char` C++ data type
This is intended to support "boxed" versions of the perlang `char` type,
i.e. a 16-bit UTF16LE character. Not currently used by generated C++
code, but will be used in an upcoming commit to emulate something like a
C# `object` on the Perlang/C++ side (for storing different kinds of
token instances in instances of a single Token class).

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/638