(Makefile) Add CppSharp binding patch infrastructure
Until now, I have been doing a few manual changes to the `PerlangCli.cs`
file after auto-generating it. When I was working on !642, I started
adding a mechanism for maintaining these changes in a more structured
way (using a `.patch` file). I extracted this change to a separate
commit here, to make the "diffs to the .patch file" from that other
branch be more readable.

(I also tried checking that the bindings are up-to-date in a CI job, but
this turned out to be problematic -
https://gitlab.perlang.org/perlang/perlang/-/merge_requests/643#note_4155)

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/644