(compiler) Fix `Token` debugger crash
Before this change, attempting to inspect a `Token` instance in the
debugger would cause a nasty crash of the Rider debugger. It turned out
this was because the `Token` class didn't handle null tokens at all. I'm
unsure this caused any problems with any "real-world" scenarios
(probably not, since we didn't have any tests failing?) but being able
to inspect values in the debugger without the whole debugging session
crashing is surely nice.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/645