(compiler) Turn some clang compiler errors into warnings instead
This was actually not _required_ in any way, but I noticed this when
working on some new functionality in the Perlang implemntation of the
`Scanner` class. Raising hard errors for private fields that are unused
feels a bit strict. Yes, it's a code smell, but this is perhaps better
handled by a linter/some "strict" mode or something. Good for
production/CI checks but not necessarily something you want to have
enabled 100% of the time while developing.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/646