(compiler) Move more `Scanner` methods to Perlang implementation
More precisely:

- `Peek`
- `PeekNext`
- `Advance`
- `IsAtEnd`

Some new methods were also added to be able to manage the `start` and
`line` fields, which is more complex now that the fields are located in
one class but utilized from another. These additional methods on the
Perlang side can be removed once the conversion is complete. The
`Scanner` class is still 782 lines long though, so we have some way to
go.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/647