(stdlib) Implement explicit bounds checking in `UTF16String`
...index operator. The bounds checking previously relied on the
`std::vector.at()` implementation, so it wasn't that it didn't exist,
but doing it ourselves gives us better control of the exact exception
message and exception type that gets thrown.

Also added a "valid range" message to the existing bounds check message
in `ASCIIString`, which could be somewhat more user friendly (since the
uppermost indexable character is 1 less than the actual size of the
string, which was the only value previously included in the error
message).

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/648