(tests) Fix some tests to not be skipped anymore
Some of these were skipped at an early stage of the "experimental
compilation" functionality. Now that native compilation is our default
mode of operation, the functionality for some of these tests is
available, so they do not need to be skipped any more.

Another bunch of tests unfortunately still need to be skipped, because
`get_type()` is not implemented for all types. I'll try to get to that
too, but in a separate follow-up commit.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/652