(compiler) Fix `get_type()` support for more expressions
Trying to do things like `(1 + 2).get_type()` didn't work. `get_type()`
is a bit of a special hack in this case, since it needs some
special-casing to be able to call a method on something which isn't
really a heap-allocated object, but just a value type. We workaround
this by just faking an appropriate
`perlang::PerlangValueTypes::get_type_*` method call.

This works to an extent in fact... and thinking more closely about it,
we can probably get quite far by doing like this. If/when a value type
gets boxed into an `object`, an object of the corresponding boxed object
type (`perlang.Int32` or whatever) will be allocated/retrieved, so
things like `object.get_type()` will work just as intended.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/653