(compiler) Add `in` expression
This was an idea I had while planning for/working on support for
`switch` statements, or rather: I thought about adding support for
ranges so that things like `case 0..9` could work, which would be quite
convenient and a lighweight form of pattern matching.

And then I thought about things like `if (i in 0..9)`, which would also
be more convenient than having to combine >= and <=. So, here goes:
support for `in` expressions for existing array types. Once we have
ranges in place, we'll support `in` for them too.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/657